requirements.txt
# Backend Dependencies
# Python 3.9+ required

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Machine Learning
scikit-learn==1.4.0

# NLP
textblob==0.17.1

# Database & ORM
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.6
pydantic==2.5.3
bcrypt==4.1.2
